Move covid dashboard pages to subdirectory
To better separate from other labs
