+++ src/pages/covid/Graph.vue
