+++ src/pages/covid/Map.vue
